Fix building System.IO.Compression.Native (#1561)
CMakeLists for S.I.Compression.Native assumes a clean room and fully configures the compiler. That's how it was written to produce clrcompression.dll.

For single file host, it was hacked to be includable from CLR build that already preconfigures the compiler in some way. Then a STATIC_LIBS_ONLY was carefully threaded through to make it sort of do what we need when included from the CLR build. But we're now in a situation where a partially configured compilation is being reconfigured and various settings are clashing.

The low impact fix is to just pass extra /guard:cf- and /GL- wherever we need. IMHO this needs to be rewritten.