Simplify integration of NonBacktracking (#1582)
Initially NonBacktracking was tacked on with its own processing loop and special-cased directly from Regex with code that detected the option and called into special functions.  I previously changed that to instead derive from RegexRunner, and added new virtual methods for it to override.  But such integration doesn't go far enough in terms of being able to plug into the currently public entry points, which will be required when we implement source generator support for NonBacktracking.

This PR fixes that up by:
- Undoing the previously added internal virtuals
- Changing SymbolicRegexRunner to instead override FindFirstChar and Go
- Deleting the custom Scan overrides
- Creating matches using the exposed Capture method rather than doing so manually

Doing this, and in particular changing it to use Capture, meant that the caps information in the Regex had to be correct, so it also:
- Fixes the Init method to properly populate these members based on the expected NonBacktracking semantics (i.e. no subcaptures)

That then required being able to pass the originally computed capture size from the pattern to Regex.Replace operations, as it uses that number to interpret whether constructs are backreferences or just text, which meant I had to store that value separately on the Regex.  And as I never like increasing the size of the object, I looked for things I could offset it with:
- Deleted the _refsInitialized member.  It was only used to record whether _replref had been initialized, which is obvious based on whether the field is non-null
- Made _replref itself initialized lazily, so that regex instances don't pay for that WeakReference object allocation unless Replace is actually being used
- That then made the old InitializeReferences protected method useless, so I made it a nop and removed the associated references