[NativeAOT-LLVM] Improve the build setup a little (#1672)
* Improve the build for NativeAOT-LLVM

1. Delete unused cmake defines from various places.
2. Move the other defines in places where they logically should be.
3. When building the CLR for Wasm, exclude everything but the nativeaot component.
4. Delete the special handling of the Jit targeting Wasm, use the standard altjit
   configuration. Add a new subset that build only this Jit, speeding up the innerloop.

* More improvements

Split the subsets for building ILC and the build integration.

This is needed because for Wasm, ILC cannot be meaningfully built,
since it needs the Jits, and is generally not useful - we will always
be cross-compiling. Up until now the ILC project included a hacky
support for cross-building, but it cannot be complete without the Jits,
and instead of making it work, I just deleted it. It will be a separate
item to enable cross-building the ILC. Note that this does not change
anything in terms of what can and cannot be done, just the artifacts
for ILC will always be used from the host's output directory, including
for the test build, which is what we want.

Tweak the YAML templates a little to delete duplicated definitions.

Tweak some CMake and build scripts to support this new model.

* Update the documentation