Change how we deal with scanning failures
When we import a method that has a problem in it (e.g. read a non-existent field because the user provided mismatched assemblies), we generate IL body for a throwing method that throws the appropriate exception matching JIT-based behavior. We then import that IL body instead of the original method.

We do this in both scanner and codegen. This worked for us so far, but I just hit a case where this doesn't work.

Before RyuJIT hits the problematic part of the method that makes us discard it, it might attempt to import stuff we didn't scan (e.g. try to do dictionary lookups, ask about virtual slots, etc.). This makes us take the "there's a compiler bug" codepaths before we find out it doesn't actually matter. We don't keep the partial dependencies from the scanning phase and even if we did, there's no guarantees RyuJIT imports in the same order as scanner - all we have is the dependencies of the throwing stub.

So instead, remember that a method had a problem during scanning and don't even attempt to give such method body to RyuJIT - just pass the throwing body right away.