Speed up access to threadstatic fields
Accessing thread statics requires calling a method in C# (because we might need to do an allocation) followed by calling a method in C++ (because that's where we keep thread-related datastructures). Both of these raise a frame and do other stuff. This inlines the fast path into assembly and tailcalls for the slow path if needed.

Speeds up accessing threadstatics by 50%.