Speed up access to threadstatic fields (#1719)
Accessing thread statics requires calling a method in C# (because we might need to do an allocation) followed by calling a method in C++ (because that's where we keep thread-related datastructures). Ideally we would want to do the inverse - call into C++ that has access to everything and only call the C# part on the slow path. But for that we need C++ to support guaranteed tailcalls and that's not going to happen anytime soon. JIT-based CoreCLR has the luxury (and curse) of being able to switch to managed world in C++ and gets away with a single call. Our only option seem to be inline assembly.

Speeds up accessing threadstatics by 50%.