Scan field RVA data blobs (#1728)
Fixes #1712.

Some RVA data blobs within the compiler are special and contain other dependencies the compiler needs to look at during scanning phase.

This fixes an issue where the `<Module>` type wasn't having its metadata generated in optimized builds because the scanning phase never saw the `<Module>` type being allocated and didn't predict it as needing metadata. The p/invoke fixup blob references the `<Module>` type as "a type from the assembly that contained the p/invoke". We need to scan the fixup blob during the scanning phase so that the type is seen.