Address feedback
Fix bugs arsing from more code coverage

The scope crept with the latest commit because removing genTreeAsLlvmType meant that some situations that failed compilation where now being attempting and some bugs were uncovered.

The math overflow helpers are all managed so they needed to be identified as such

The return arg and this arg need to be identifed when creating the call as the LLVM signature is built from the `CORINFO_SIG_INFO`.  I expect this will get cleared up when, or after, the PUTARG work is done.

Ctors returning `this` need to be excluded currently for the same reason: the CORINFO_SIG_INFO shows those methods as `void`.

There seems to be a difference in phi rules between IR and LLVM (https://github.com/dotnet/runtimelab/issues/1731) so I removed the restriction on the local only appearing once in the phi args.