[NativeAOT-LLVM] Enable IL scanner for LLVM to enable precomputed vtable slots (#1834)
This PR enables the IL scanner for LLVM.  The scanner is always on for LLVM to enable precomputed vtables slots.  

This will unblock the `CT_INDIRECT` work https://github.com/dotnet/runtimelab/pull/1768 .  At ComputeGenericLookup in `Compilation.cs`  you can see that it never takes the fixed lookup path for generic calls, so not sure that if that is going to be a problem.

There are some changes to the scanner to add dependencies required by the IL->LLVM compilation which should be removed when all code is generated by clrjit. Including

- The scanner has the same stubbing out of methods not implemented for LLVM, `GetRandomBytes` and `EnumCalendarInfo` 
- For expanding intrinsics, the scanner has to add dependencies for both expanding and not expanding as it doesn't know which module will compile the method.
- Adding dependencies for throwing exceptions and exception handling
- Various other places where the IL->LLVM compilation differs from the scanner