Validation for RuntimeFunctions and change to unwindinfo (#1960)
* Validation for RuntimeFunctions and change to unwindinfo

Added a function to validate the information for RuntimeFunctions via asserts that make sure they're correctly sorted, their intervals don't overlap, and their PersonalityRoutineRVA is the same. Also modified UnwindInfo so that whenever the CHAININFO flag is on we don't generate the personality routine, as it was showing the RVA to the hot function instead of the Personality Routine.

* Update to ValidateRuntimeFunctions

Updated ValidateRuntimeFunctions to account for the possibility of up to two different PersonalityRoutineRVAs, as there can be one for filter funclet and one for the other things.