[NativeAOT-LLVM] Exception handling: funclets (#2109)
* Place locals live-in/out into/from handlers on the shadow stack

Requires reworking our model to not assume that if a given local
is live on the shadow stack at all, it is live on it always. We
will need this in the future anyway, for optimizations.

* Support for generating funclets

With these changes we can successfully (i. e. without asserts) generate
functions (of arbitrary complexity) with EH. Of course, the actual EH
dispatch part is NYI.

Also, debugging in fuclets is broken.

* Disable compiling synchronized methods