Pass the original shadow stack pointer to funclets
Consider the first pass of exception handling. The dispatcher
will call filter funclets, which can invoke arbitrary managed
code. This means filters can invalidate all of the shadow state
below them, even as that state is still live (recall we're only
in the first pass).

Consider the second pass of exception handling. The dispatcher
is sitting, in the shadow call stack, below the finally and fault
handlers, therefore, they can overwrite the shadow state of the
dispatcher itself.

Consider the exception object. It must live on the shadow stack,
and is a part of the dispatcher's shadow state (allocated below
the "throw" point). Therefore, it is not sufficient to simply
ensure the dispatcher itself holds no live shadow state between
the calls to funclets to solve the problem above.

What is sufficient is using a dedicated shadow stack pointer
to access the original shadow frame in funclets. It is how the
other backends operate as well. This also solves the issue of
passing the exception object to funclets nicely and naturally:
it is allocated and passed as the first shadow parameter.

This solution has the drawback of not "being optimal" for the
case of locally invoked finallys. We presume that to be fine,
as do the other backends.