Revert "Testing..."
Adding this made the bug go away; this is a data point.

The error here is that we're calling a non-existent function pointer in
"S_P_CoreLib_Internal_Reflection_Extensions_NonPortable_CustomAttributeSearcher_1__GetMatchingCustomAttributesIterator_d__2<System___Canon>__MoveNext".

This is a method with (native) EH, so all potentially throwing calls are rountripped through JS by Emscripten.
Therefore, we are looking for an invoke which could potentially have a corrupted target.

We assume native stack corruption is exceedingly unlikely to be the cause, therefore our primary targets are
indirect calls made through code generated by us, with function pointers potentially dependent on shadow state
(recall we are assuming the underlying cause to be shadow stack corruption). Virtual (including interface) calls
fit this description.

Data point: the method in question is compiled by the IL backend, which is known to have shadow stack corrupting
bugs for methods with EH, though I am not sure of their exact nature.

It would be great to identify the exact crash site. For now though, just add logging of compilation order...