[NativeAOT-LLVM] Fix most warnings for the `HelloWasm` runtime test (#2168)
* Fix tentative instance method target

There are two throw helpers used for tentative methods: "ThrowBodyRemoved"
and "ThrowInstanceBodyRemoved". We were always using the first one.

* Add DAM to silence linker warnings

* Disable implicit SDL headers

* Prevent tests from importing "normal" NuGet props/targets

We import them manually from a shared location.

Usually this is not a problem because batch build (via the build script)
does its own restore thing, but it is a problem in case the project is
restored on its own, such as when it is built with "dotnet build".