NativeAOT-LLVM: merge PR 67238 - JIT: Refactor call argument representation (#2177)
* JIT: Refactor call argument representation (#67238)

This refactors the JIT's representation of call arguments. It replaces
`GenTreeCall::Use` and `fgArgTabEntry` with a single class `CallArg`.
`CallArg` always contains space for ABI information and contains two
intrusive linked list nodes: one for all args (similar to current
`gtCallArgs`) where all standard arguments are always in argument order,
and one for late args (similar to current `gtCallLateArgs`) that may be
reordered. The late args list may also not contain all arguments.

`fgArgInfo` is also replaced by a new class `CallArgs` that is stored
inline in `GenTreeCall`. It encapsulates all handling of arguments
(insertion/removal). The change also begins treating the 'this' argument
as a normal argument with `CallArgs` providing convenient access to it
when necessary.

The main benefit of this change is to avoid keeping track of the side
table `fgArgInfo` and having to scan through this side table repeatedly
when we need to query argument information. In addition it gives more
convenient ways to access well known arguments like 'this', the ret
buffer, VSD cell, etc. Finally, it also serves as a nice clean-up in the JIT.

* trivially resolve conflicts

* put back closing endif

* Clear errors in gentree.cpp:

- remove customregister cases for Wasm for R2RIndirectionCell and ValidateIndirectCallTarget - might need revisiting as not sure we will hit these.
- temporary function for arg message for gtDispLIRNode with GT_PUTARG_TYPE

* remove non llvm file compilation errors

- Remove compiler.codegen field
- Change assert for HasRetBufArg - to iciCall->gtArgs.HasRetBuffer() - might have to revisit this

* remove GenTreeCall

* define MIN_ARG_AREA_FOR_CALL for LLVM/Wasm

* lower calls using gtArgs

* whitespace cleanup

* address feedback, move things around, add comments, delete dead code.

* Update src/coreclr/jit/llvmlower.cpp

Co-authored-by: SingleAccretion <62474226+SingleAccretion@users.noreply.github.com>

* Update src/coreclr/jit/compiler.h

Co-authored-by: SingleAccretion <62474226+SingleAccretion@users.noreply.github.com>

* Update src/coreclr/jit/gentree.cpp

Co-authored-by: SingleAccretion <62474226+SingleAccretion@users.noreply.github.com>

* Update src/coreclr/jit/llvmlower.cpp

Co-authored-by: SingleAccretion <62474226+SingleAccretion@users.noreply.github.com>

* RemovedWellKnownArg for head removal

* add shadow stack and return slot in the order they appear

* simplify getting first arg.

---------

Co-authored-by: Jakob Botsch Nielsen <Jakob.botsch.nielsen@gmail.com>
Co-authored-by: SingleAccretion <62474226+SingleAccretion@users.noreply.github.com>