[NativeAOT-LLVM] Implement more things (#2212)
* Disable fast tail calls

We should definitely implement these, but they require ABI
changes which should go hand-in-hand with the fixes for the
native ABI, i. e. after the IL backend is gone.

* Delete non-sig arg handling for helpers

The only helper which could hit this path is not used in NAOT.

* Delete the "hasExplicitThis" block

VM will never pass us this flag, and and the Jit doesn't
actually handle it properly for the methods being compiled.

* Normalize struct call uses

* Implement function pointer symbols

Plus a few misc fixed.

Sharing the signature building code was deferred to the ABI fix change.

* Delete ReplaceStubbedWasmMethods

Was causing scanning failures becase we didn't scan the UCO method originating from "ldftn".

* Delete TYP_STRUCT from getLlvmTypeForVarType

* Implement GT_MEMORYBARRIER

* Define shadow stack top in the runtime