Do not designate the RPI helper as HFIF_NO_RPI_OR_GC
The helper may invoke runtime initialization, which runs managed code.
Not saving the shadow stack beforehand can lead to various kinds of
hard-to-debug issues. For example, one failure mode was that we would
try to allocate two shadow stacks, leading later to a GC hole because
the shadow stack bottom would be above its top.