[NativeAOT-LLVM] More miscellaneous fixes (#2388)
* Disable background GC support

Turns out it was enabled all along!

* Stub out more threading

* Add some notes about exports/imports

* A couple more doc updates

* Add some linker options

1) "--strip-all" so that we strip both DWARF and the names section
   if the user specifies "/p:NativeDebugSymbols=false".
2) Also compress relocations if we are throwing away the debug info.
   Recall this saves ~8% in terms of code size.
2) Fix the global base and stack size.

* Add a note about NativeDebugSymbols to the doc

* Do not designate the RPI helper as HFIF_NO_RPI_OR_GC

The helper may invoke runtime initialization, which runs managed code.
Not saving the shadow stack beforehand can lead to various kinds of
hard-to-debug issues. For example, one failure mode was that we would
try to allocate two shadow stacks, leading later to a GC hole because
the shadow stack bottom would be above its top.