Optimize explicit frame release from unwinding faults
When we are unwinding past a virtual frame, we must unlink it.
For top-level faults this is done via an explicit helper call
in codegen and we assumed that all faults could use the virtual
unwind frame of their parent function.

This change refines this so that only faults with nested catch
handlers need to do this while for others, the responsibility
for releasing the frame is "pushed down" to the inner handler.

This is achieved by giving top-level EH regions protected by
faults NOT_IN_TRY unwind index instead of NOT_IN_TRY_CATCH.

An example of where this optimization helps is when we have
the common "try { } catch { } finally { }" pattern in C#, which
is represented as "try { try { } catch { } } finally { }" in IL.