huge hack: LOCAL_VAR is not usable with split stack
The LOCAL_VAR interpreter macro assumes that ref and non-ref parts of
a `stackval*` are at the same offset and uses a cast through a C
pointer type (either MonoObject* or a value type) to acess the lvalue
for a local.

But with the split stack those are different offsets.

Change the LOCAL_VAR macro to access the valuetype portion.
In the future we will need to audit all uses of LOCAL_VAR and go
through some other macro (LOCAL_VAR_O) to access the object part.

Also in the case where the value is accessed as a `stackval*`, add a
new LOCAL_VAR_SV macro.