Add an initial implementation of the LSSA testing scheme
This allows us to write validation tests such as the following:

[LSSATestAttribute("""
 BB01:
   ZEROINIT 4
   STORE SS00 ARG00 ARG00/1
   STORE SS01 TMP00 TMP00/1
 """)]
private static void Problem(object a)
{
    object x = new object();
    Console.WriteLine(a);
    Console.WriteLine(a);
    Console.WriteLine(x);
    Console.WriteLine(x);
    return;
}

To make the scheme robust against non-essential IR differences, the validation
is performed on a small intermediate representation - an array of "allocation
actions", which can both be collected from the IR and parsed out from text.

For now this is enabled for non-DEBUG builds, since we test with a Release
compiler in CI, but ideally, this should only be compiled in Debug/Checked.