exit `DispatchWasiEventLoop` without polling if tasks have been canceled
If one or more tasks have been canceled during the call to
`ThreadPoolWorkQueue.Dispatch`, one or more tasks of interest to the application
may have completed, so we return control immediately without polling, allowing
the app to exit if it chooses.

A practical example of this is in the SharedLibrary smoke test.  Without this
patch, that test will take over 100 seconds to complete, whereas with this patch
it completes in under a second.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>