Add support for async struct instance methods
This adds support for async2 struct instance calls, with the restriction
that the "this" argument passed to these calls in IL must be
syntactically either `ldloca <some local>` or `ldarg 0`, with the latter
only being allowed for self-instance calls. Furthermore, async2 struct
instance methods are not allowed to use the `ldarga 0` or `starg 0`
instructions.

Conceptually the VM and JIT rewrite these calls from "await x.Foo()"
into something corresponding to "x = Foo(x)". Technically, the way this
works is that on suspension, a struct-instance call will box its own
instance such that it can restore it again on resumption. Additionally,
during resumption we propagate the struct instance back to the caller
continuation as if it was a return result: for self-instance calls
resumption is done simply by overriding the "this" pointer from the
callee, while the local calls resumption is done by overwriting the
local on the stack frame.