Clean up external method cell code
Address a TODO about not passing the signature around.

Remove a TODO about using getAddressOfPInvokeTarget.
Fundamentally, the indirection is needed for LLVM
reasons, NOT WASM reasons (with straight-line WASM
we could simply emit i32.const <sym> call_indirect),
and it's cleaner to have this separation between LLVM
and not LLVM how it is right now, a separate method.

This a no-diff change.